update operator and docs for LU factorization
