+++ test/mlu_op_gtest/pb_gtest/src/zoo/sgetrf/test_case/case0.prototxt
@@ -2,8 +2,8 @@ op_name: "test_sgetrf"
 input {
   id: "input"
   shape: {
-    dims: 256
-    dims: 128
+    dims: 1024
+    dims: 1024
  
   }
   layout: LAYOUT_ARRAY
@@ -18,8 +18,8 @@ input {
 output {
   id: "output"
   shape: {
-    dims: 256
-    dims: 128
+    dims: 1024
+    dims: 1024
    
   }
   layout: LAYOUT_ARRAY
@@ -28,7 +28,7 @@ output {
 test_param: {
   error_func: DIFF1
   error_func: DIFF2
-  error_threshold: 0.000003
-  error_threshold: 0.000003
+  error_threshold: 0.003
+  error_threshold: 0.003
   baseline_device: CPU
 }
